+++ x-pack/plugin/relevance-search/src/main/java/org/elasticsearch/xpack/relevancesearch/RelevanceSearchPlugin.java
@@ -41,11 +41,11 @@ import org.elasticsearch.xpack.relevancesearch.query.RelevanceMatchQueryRewriter
 import org.elasticsearch.xpack.relevancesearch.settings.curations.CurationsService;
 import org.elasticsearch.xpack.relevancesearch.settings.index.IndexCreationService;
 import org.elasticsearch.xpack.relevancesearch.settings.relevance.RelevanceSettingsService;
-import org.elasticsearch.xpack.relevancesearch.xsearch.XSearchAnalyticsService;
 import org.elasticsearch.xpack.relevancesearch.xsearch.XSearchRequestValidationService;
 import org.elasticsearch.xpack.relevancesearch.xsearch.action.XSearchAction;
 import org.elasticsearch.xpack.relevancesearch.xsearch.action.XSearchTransportAction;
 import org.elasticsearch.xpack.relevancesearch.xsearch.action.rest.RestXSearchAction;
+import org.elasticsearch.xpack.relevancesearch.xsearch.analytics.XSearchAnalyticsService;
 
 import java.util.Collection;
 import java.util.Collections;
